// src/app/shared/util-model/enumeration/event-profile-action.enum.ts
export enum EventProfileActionEnum {
    SELECT_EVENT_PROFILE = '[Backend] Selecting event\'s profile',
    UPDATE_EVENT_PROFILE = '[Backend] Updating event\'s profile',
    BLOCK_EVENT_PROFILE = '[Backend] Blocking event\'s profile',
    UNBLOCK_EVENT_PROFILE = '[Backend] Unblocking event\'s profile',
    DELETE_EVENT_PROFILE = '[Backend] Deleting event\'s profile',
}